Move rich text native to separate package (#16219)
* Move React Native RichText to new package

* Fix multiline reading.

* Move default styles to component.

* Move tests to component.

* Fix paste on title
