+++ packages/block-editor/src/components/rich-text/format-toolbar/index.native.js
@@ -4,10 +4,10 @@
 
 import { Toolbar, Slot } from '@wordpress/components';
 
-const FormatToolbar = ( { controls } ) => {
+const FormatToolbar = () => {
 	return (
 		<Toolbar>
-			{ controls.map( ( format ) =>
+			{ [ 'bold', 'italic', 'link' ].map( ( format ) =>
 				<Slot name={ `RichText.ToolbarControls.${ format }` } key={ format } />
 			) }
 			<Slot name="RichText.ToolbarControls" />
