+++ packages/editor/src/components/post-title/style.native.scss
@@ -6,3 +6,4 @@
 	padding-bottom: $title-block-padding-bottom;
 	margin-top: 24;
 }
+
