+++ packages/rich-text/src/component/style.native.scss
@@ -1,11 +1,11 @@
 
-.block-editor-rich-text {
+.rich-text {
 	font-family: $default-regular-font;
 	min-height: $min-height-paragraph;
 	color: $gray-900;
 	text-decoration-color: $blue-500;
 }
 
-.block-editor-rich-text-placeholder {
+.rich-text-placeholder {
 	color: $gray;
 }
