+++ packages/rich-text/src/component/test/index.native.js
